requirements.txt
python-binance==0.7.10
schedule==1.0.0
apprise==0.9.1
